Move _CaseInfo to its own module
